// src/Enums/AppModal.ts
export enum CharacterNames {
  Name = 'characterName',
  Status = 'characterStatus',
  Species = 'characterSpecies',
  Type = 'characterType',
  Gender = 'characterGender'
}

export enum CharacterPlaceholders {
  Name = 'Add Name',
  Status = 'Add Status',
  Species = 'Add Species',
  Type = 'Add Type',
  Gender = 'Add Gender'
}